requirements.txt
Django
djangorestframework
psycopg2
drf-yasg